🎨 chore: satisfy swiftlint — split exporter tests + suppress view body length
- `ResultMarkdownExporterTests.swift` hit the 400-line `file_length` cap
  after the #143 phase-path tests landed. Split the 3 new tests into
  `ResultMarkdownExporterTests+PhasePath.swift` per the
  "Splitting a Suite Across Files" rule (sibling extension, not a new
  `@Suite`, so they inherit the original's `.serialized` /
  `.timeLimit` / `@MainActor` traits). Fixture helpers dropped to
  internal access so the extension can reach them. Identifier-name
  violations (`tl`/`sp`/`a`/`b`) renamed to `top`/`sub`/`alice`/`bob`
  while moving.

- `ResultDetailView.swift` body length crept from 239 → 260 lines
  with the `subPhaseWrapper` addition. Add a `type_body_length`
  disable comment — the existing body has organic reasons for being
  long (onAppear loader, multiple row builders, navigation wiring) and
  splitting it out would pull interlocking `@State` references across
  a boundary.
