+++ Pastura/PasturaTests/App/ResultMarkdownExporterTests.swift
@@ -6,12 +6,17 @@ import Testing
 @Suite(.timeLimit(.minutes(1))) @MainActor struct ResultMarkdownExporterTests {  // swiftlint:disable:this type_body_length
 
   // MARK: - Fixtures
+  //
+  // Helpers are at internal access (not `private`) so the
+  // `ResultMarkdownExporterTests+PhasePath.swift` sibling extension can reuse
+  // them. `private` members aren't visible to cross-file extensions — see
+  // `.claude/rules/testing.md` "Splitting a Suite Across Files".
 
-  private let createdAt = Date(timeIntervalSince1970: 1_712_000_000)  // 2024-04-01T19:33:20Z
-  private let updatedAt = Date(timeIntervalSince1970: 1_712_000_342)  // +5m 42s
-  private let exportAt = Date(timeIntervalSince1970: 1_713_000_000)  // 2024-04-13T08:53:20Z
+  let createdAt = Date(timeIntervalSince1970: 1_712_000_000)  // 2024-04-01T19:33:20Z
+  let updatedAt = Date(timeIntervalSince1970: 1_712_000_342)  // +5m 42s
+  let exportAt = Date(timeIntervalSince1970: 1_713_000_000)  // 2024-04-13T08:53:20Z
 
-  private func makeScenario(
+  func makeScenario(
     id: String = "s1",
     name: String = "Prisoners Dilemma",
     yaml: String = "name: Prisoners Dilemma\nrounds: 2\n"
@@ -21,7 +26,7 @@ import Testing
       isPreset: true, createdAt: Date(), updatedAt: Date())
   }
 
-  private func makeSimulation(
+  func makeSimulation(
     id: String = "sim1",
     scenarioId: String = "s1",
     status: SimulationStatus = .completed,
@@ -37,7 +42,7 @@ import Testing
       modelIdentifier: modelIdentifier, llmBackend: llmBackend)
   }
 
-  private func makeTurn(
+  func makeTurn(
     round: Int,
     seq: Int,
     phase: String,
@@ -62,7 +67,7 @@ import Testing
       createdAt: Date())
   }
 
-  private func makeState(
+  func makeState(
     scores: [String: Int] = ["Alice": 5, "Bob": 3],
     eliminated: [String: Bool] = [:]
   ) -> SimulationState {
@@ -77,7 +82,7 @@ import Testing
       currentRound: 2)
   }
 
-  private func makeExporter(
+  func makeExporter(
     filter: ContentFilter = ContentFilter(blockedPatterns: [])
   ) -> ResultMarkdownExporter {
     ResultMarkdownExporter(
@@ -354,102 +359,7 @@ import Testing
   }
 
   // MARK: - Phase path grouping
-
-  @Test func nestedAndTopLevelSamePhaseTypeProduceTwoDistinctHeadings() throws {
-    // path [0] → top-level, path [1,0] → nested sub-phase; same phaseType "speak_all"
-    // must produce two separate headings rather than collapsing into one.
-    let exporter = makeExporter()
-    let topLevel = makeTurn(
-      round: 1, seq: 1, phase: "speak_all",
-      agent: "Alice", fields: ["statement": "top level"],
-      phasePathJSON: "[0]")
-    let nested = makeTurn(
-      round: 1, seq: 2, phase: "speak_all",
-      agent: "Bob", fields: ["statement": "nested sub-phase"],
-      phasePathJSON: "[1,0]")
-    let input = ResultMarkdownExporter.Input(
-      simulation: makeSimulation(),
-      scenario: makeScenario(),
-      turns: [topLevel, nested],
-      state: makeState())
-
-    let result = try exporter.export(input)
-
-    #expect(result.text.contains("#### Phase: speak_all"))
-    #expect(result.text.contains("#### Sub-phase: speak_all (path [1, 0])"))
-    // Alice belongs under the top-level heading, Bob under the sub-phase heading.
-    let topLevelRange = result.text.range(of: "#### Phase: speak_all")
-    let subPhaseRange = result.text.range(of: "#### Sub-phase: speak_all (path [1, 0])")
-    let aliceRange = result.text.range(of: "**Alice**")
-    let bobRange = result.text.range(of: "**Bob**")
-    #expect(topLevelRange != nil && subPhaseRange != nil)
-    #expect(aliceRange != nil && bobRange != nil)
-    // Top-level heading should appear before the sub-phase heading (first-seen order).
-    if let tl = topLevelRange, let sp = subPhaseRange {
-      #expect(tl.lowerBound < sp.lowerBound)
-    }
-    // Alice should appear before the sub-phase heading (she's in the top-level block).
-    if let a = aliceRange, let sp = subPhaseRange {
-      #expect(a.lowerBound < sp.lowerBound)
-    }
-    // Bob should appear after the sub-phase heading.
-    if let b = bobRange, let sp = subPhaseRange {
-      #expect(b.lowerBound > sp.lowerBound)
-    }
-  }
-
-  @Test func mixedEraLegacyAndTopLevelSamePhaseTypeGroupTogether() throws {
-    // Legacy (nil path) and v6 top-level ([0]) for the same phaseType must
-    // render under a single "#### Phase: speak_all" heading in sequence order.
-    let exporter = makeExporter()
-    let legacy = makeTurn(
-      round: 1, seq: 1, phase: "speak_all",
-      agent: "Alice", fields: ["statement": "legacy turn"],
-      phasePathJSON: nil)
-    let newTopLevel = makeTurn(
-      round: 1, seq: 2, phase: "speak_all",
-      agent: "Bob", fields: ["statement": "v6 turn"],
-      phasePathJSON: "[0]")
-    let input = ResultMarkdownExporter.Input(
-      simulation: makeSimulation(),
-      scenario: makeScenario(),
-      turns: [legacy, newTopLevel],
-      state: makeState())
-
-    let result = try exporter.export(input)
-
-    // Exactly one top-level heading for speak_all — not two.
-    let occurrences = result.text.components(separatedBy: "#### Phase: speak_all").count - 1
-    #expect(occurrences == 1)
-    // Both agents appear; Alice (legacy, seq=1) before Bob (v6, seq=2).
-    let aliceRange = result.text.range(of: "**Alice**")
-    let bobRange = result.text.range(of: "**Bob**")
-    #expect(aliceRange != nil && bobRange != nil)
-    if let a = aliceRange, let b = bobRange {
-      #expect(a.lowerBound < b.lowerBound)
-    }
-    // No sub-phase heading should appear.
-    #expect(!result.text.contains("#### Sub-phase:"))
-  }
-
-  @Test func orphanSubPhaseRendersWithoutParentHeading() throws {
-    // A conditional sub-phase turn (path [0,0]) without a top-level parent
-    // persisted must render as "#### Sub-phase: speak_all (path [0, 0])".
-    // No "#### Phase: speak_all" heading is expected.
-    let exporter = makeExporter()
-    let subPhaseTurn = makeTurn(
-      round: 1, seq: 1, phase: "speak_all",
-      agent: "Alice", fields: ["statement": "from sub-phase"],
-      phasePathJSON: "[0,0]")
-    let input = ResultMarkdownExporter.Input(
-      simulation: makeSimulation(),
-      scenario: makeScenario(),
-      turns: [subPhaseTurn],
-      state: makeState())
-
-    let result = try exporter.export(input)
-
-    #expect(result.text.contains("#### Sub-phase: speak_all (path [0, 0])"))
-    #expect(!result.text.contains("#### Phase: speak_all"))
-  }
+  //
+  // Split into sibling file `ResultMarkdownExporterTests+PhasePath.swift` to
+  // stay under the 400-line file_length cap. See `.claude/rules/testing.md`.
 }
